Split the shift key_collection module, drop its dead API surface
The shift protocol's key_collection.rs had grown to 1090 lines mixing
six unrelated concerns: the Operation enum, the dense shift-sequence
encoding, Key/ConstraintIndex, KeySegment, KeyCollection, and the
constraint-system-walking builder, plus all their serialization impls
and tests in one file.

An exhaustive grep across the whole workspace (the only reliable check
here, since rustc's dead_code/unreachable_pub lints can't see "pub
method whose type is reachable but nobody actually calls it") turned up
real dead weight: KeyCollection::n_words() had zero callers anywhere,
including its own tests, and DenseShiftEncoding::decode() and
Key::accumulate_by_operand() had no production callers, only exercised
by this file's own tests.

This splits the file into one module per concern (operation.rs,
dense_shift_encoding.rs, key.rs, key_segment.rs, collection.rs,
builder.rs, mod.rs), each with its own tests, and:

- deletes KeyCollection::n_words() (unused everywhere)
- drops decode() entirely, inlining its one real use in a test
- moves accumulate_by_operand() into key.rs's own test module as a
  plain test helper instead of a #[cfg(test)] method on Key
- drops the inert #[repr(u16)] on Operation (the wire format is defined
  independently via explicit match arms, so the attribute did nothing)
- turns the free update_with_operand/update_with_constraints functions
  into methods on a small BuilderKeyLists newtype, so
  build_key_collection reads as a sequence of actions on one subject
  instead of threading a mutable slice through every call
- points m4-prover's build_g_from_folded_words at the existing
  word_keys()/n_words() accessors instead of re-deriving the same slice
  by hand

Pure refactor plus dead-code removal, no behavior change. Verified with
cargo check --workspace --all-targets, clippy -D warnings --all-targets
and cargo doc -D warnings on binius-prover/binius-m4-prover, and the
full shift-protocol test suite (30 unit + prove_verify + shift
integration + 2 in m4-prover), all passing.
